# greet-name-service/openapi.yaml
openapi: 3.0.0
info:
  title: Greeter API
  description: This API provides a greeter service that interacts with two services using OAuth2 authentication.
  version: 1.1.0

servers:
  - url: http://localhost:8080
    description: Local server

paths:
  /greeter/greet:
    get:
      summary: Get a greeting message from the greeter
      description: Makes a project-level request to retrieve a greeting message.
      responses:
        '200':
          description: Successfully returns a greeting message from the greeter.
          content:
            application/json:
              schema:
                type: string
                example: "Hello, person!"
        '500':
          description: Error occurred while processing the request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /greeter/greetDb:
    get:
      summary: Check DB connection
      description: hehe
      responses:
        '200':
          description: Successfully returns env variables
          content:
            application/json:
              schema:
                type: string
                example: "Hello, person!"
        '500':
          description: Error occurred while processing the request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /greeter/greetDbProd:
    get:
      summary: Check DB connection
      description: hehe
      responses:
        '200':
          description: Successfully returns env variables
          content:
            application/json:
              schema:
                type: string
                example: "Hello, person!"
        '500':
          description: Error occurred while processing the request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /greeter/greetOrg:
    get:
      summary: Get a greeting message from the greeter
      description: Makes an OAuth2 authenticated org-level request to return a greeting message.
      responses:
        '200':
          description: Successfully returns a greeting message from the greeter.
          content:
            application/json:
              schema:
                type: string
                example: "Hello, person!"
        '500':
          description: Error occurred while processing the request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - oauth2_client_credentials: []

components:
  securitySchemes:
    oauth2_client_credentials:
      type: oauth2
      description: OAuth2 client credentials flow for the org-level service
      flows:
        clientCredentials:
          tokenUrl: https://sts.preview-dv.choreo.dev/oauth2/token
          scopes: {}

  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
      example:
        message: "Internal server error"